agent_workspace/requirements.txt
# Core agent framework
agno>=0.9.0

# Perplexity model + search tool (bundled with agno, listed for clarity)
# Ensure your agno version includes agno.models.perplexity and agno.tools.perplexity

# Environment management
python-dotenv>=1.0.0

# Optional: persistent vector memory
psycopg2-binary>=2.9.9
pgvector>=0.2.4

# Optional: structured output / JSON parsing safety
pydantic>=2.0.0
